+++ compiler-cli/src/lib.rs
@@ -74,6 +74,7 @@ mod publish;
 mod remove;
 pub mod run;
 mod shell;
+mod text_layout;
 
 use config::root_config;
 use fs::{get_current_directory, get_project_root};
@@ -403,6 +404,9 @@ enum Dependencies {
     /// Download all dependency packages
     Download,
 
+    /// List all outdated dependencies
+    Outdated,
+
     /// Update dependency packages to their latest versions
     Update(UpdateOptions),
 
@@ -585,6 +589,11 @@ fn parse_and_run_command() -> Result<(), Error> {
         Command::Deps(Dependencies::Download) => {
             let paths = find_project_paths()?;
             download_dependencies(&paths)
+        }
+
+        Command::Deps(Dependencies::Outdated) => {
+            let paths = find_project_paths()?;
+            dependencies::outdated(&paths)
         }
 
         Command::Deps(Dependencies::Update(options)) => {