+++ compiler-core/src/type_/error.rs
@@ -6,6 +6,7 @@ use crate::{
     ast::{BinOp, BitArraySegmentTruncation, Layer, SrcSpan, TodoKind},
     build::Target,
     exhaustiveness::ImpossibleBitArraySegmentPattern,
+    parse::LiteralFloatValue,
     type_::{Type, expression::ComparisonOutcome},
 };
 
@@ -671,6 +672,10 @@ pub enum Error {
     ExternalTypeWithConstructors {
         location: SrcSpan,
     },
+
+    LowercaseBoolPattern {
+        location: SrcSpan,
+    },
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -1309,7 +1314,8 @@ impl Error {
             | Error::NonUtf8StringAssignmentInBitArray { location }
             | Error::PrivateOpaqueType { location }
             | Error::SrcImportingDevDependency { location, .. }
-            | Error::ExternalTypeWithConstructors { location, .. } => location.start,
+            | Error::ExternalTypeWithConstructors { location, .. }
+            | Error::LowercaseBoolPattern { location } => location.start,
             Error::UnknownLabels { unknown, .. } => {
                 unknown.iter().map(|(_, s)| s.start).min().unwrap_or(0)
             }
@@ -1978,17 +1984,14 @@ pub fn check_javascript_int_safety(int_value: &BigInt, location: SrcSpan, proble
 /// Erlang's floating point numbers
 ///
 pub fn check_erlang_float_safety(
-    string_value: &EcoString,
+    value: LiteralFloatValue,
     location: SrcSpan,
     problems: &mut Problems,
 ) {
     let erl_min_float = -1.7976931348623157e308f64;
     let erl_max_float = 1.7976931348623157e308f64;
 
-    let float_value: f64 = string_value
-        .replace("_", "")
-        .parse()
-        .expect("Unable to parse string to floating point value");
+    let float_value = value.value();
 
     if float_value < erl_min_float || float_value > erl_max_float {
         problems.error(Error::ErlangFloatUnsafe { location });