+++ compiler-core/src/type_/expression.rs
@@ -16,7 +16,7 @@ use crate::{
     },
     build::Target,
     exhaustiveness::{self, CompileCaseResult, CompiledCase, Reachability},
-    parse::PatternPosition,
+    parse::{LiteralFloatValue, PatternPosition},
     reference::ReferenceKind,
 };
 use ecow::eco_format;
@@ -454,15 +454,17 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
             } => Ok(self.infer_tuple(elements, location)),
 
             UntypedExpr::Float {
-                location, value, ..
+                location,
+                value,
+                float_value,
             } => {
                 if self.environment.target == Target::Erlang
                     && !self.current_function_definition.has_erlang_external
                 {
-                    check_erlang_float_safety(&value, location, self.problems)
+                    check_erlang_float_safety(float_value, location, self.problems)
                 }
 
-                Ok(self.infer_float(value, location))
+                Ok(self.infer_float(value, float_value, location))
             }
 
             UntypedExpr::String {
@@ -665,10 +667,16 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
         }
     }
 
-    fn infer_float(&mut self, value: EcoString, location: SrcSpan) -> TypedExpr {
+    fn infer_float(
+        &mut self,
+        value: EcoString,
+        float_value: LiteralFloatValue,
+        location: SrcSpan,
+    ) -> TypedExpr {
         TypedExpr::Float {
             location,
             value,
+            float_value,
             type_: float(),
         }
     }
@@ -1955,6 +1963,22 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
                     _ => return,
                 }
             }
+
+            (
+                TypedExpr::Float { float_value: n, .. },
+                op,
+                TypedExpr::Float { float_value: m, .. },
+            ) => match op {
+                BinOp::LtFloat if n < m => ComparisonOutcome::AlwaysSucceeds,
+                BinOp::LtFloat => ComparisonOutcome::AlwaysFails,
+                BinOp::LtEqFloat if n <= m => ComparisonOutcome::AlwaysSucceeds,
+                BinOp::LtEqFloat => ComparisonOutcome::AlwaysFails,
+                BinOp::GtFloat if n > m => ComparisonOutcome::AlwaysSucceeds,
+                BinOp::GtFloat => ComparisonOutcome::AlwaysFails,
+                BinOp::GtEqFloat if n >= m => ComparisonOutcome::AlwaysSucceeds,
+                BinOp::GtEqFloat => ComparisonOutcome::AlwaysFails,
+                _ => return,
+            },
 
             _ => return,
         };
@@ -3773,13 +3797,19 @@ impl<'a, 'b> ExprTyper<'a, 'b> {
             }
 
             Constant::Float {
-                location, value, ..
+                location,
+                value,
+                float_value,
             } => {
                 if self.environment.target == Target::Erlang {
-                    check_erlang_float_safety(&value, location, self.problems)
+                    check_erlang_float_safety(float_value, location, self.problems)
                 }
 
-                Ok(Constant::Float { location, value })
+                Ok(Constant::Float {
+                    location,
+                    value,
+                    float_value,
+                })
             }
 
             Constant::String {
@@ -5208,7 +5238,7 @@ fn static_compare(one: &TypedExpr, other: &TypedExpr) -> StaticComparison {
             }
         }
 
-        (TypedExpr::Float { value: n, .. }, TypedExpr::Float { value: m, .. }) => {
+        (TypedExpr::Float { float_value: n, .. }, TypedExpr::Float { float_value: m, .. }) => {
             if n == m {
                 StaticComparison::CertainlyEqual
             } else {